Break up pick_dataset into smaller classes by dependency
